remove support for congruence closure computation
Supporting congruence closure requires the `unioned_keys` iterator,
which in turn requires maintaining the parent/sibling pointers. This
grows the size of nodes, which in turn leads to a 2% slowdown in
various rustc test cases (e.g., the `coercions` benchmark in
perf.rust-lang.org). Given that this feature is not used in chalk or
anywhere else, no reason to support it.